Created @vocdoni/react-providers
I've removed vite in favor of tsup because I've not been able to make
work the bundles generated by vite properly (all I tried was causing
context errors).

But everything related to bundling now works better:
- Now ts files are used as entry points, meaning you won't need to run
  `yarn build` in a package when developing and trying stuff in other
  packages.
- These ts files are now replaced on prepack using the clean-package
  package.
- The resulting npmjs package ends up being the same as it's now, a
  properly bundled bundle of js files

I also had to reconsider how we were treating dependencies in some
packages. The functionality is the same, we'll need to install
react-markdown and its dependency because it's left as peerDependency.
This is done this way because in the future we should try to set this
dependency as optional (only enable markdown rendering if it is
installed and detected).

rainbow-wallets should also be updated to use tsup, so it follows the
same principles as the other packages, and also so we can import it
without bundling issues.

refs #79
