+++ packages/chakra-components/src/components/Election/Description.tsx
@@ -1,8 +1,7 @@
 import { ChakraProps, useStyleConfig } from '@chakra-ui/system'
+import { useElection } from '@vocdoni/react-providers'
 import { ReactMarkdownProps } from 'react-markdown/lib/complex-types'
-
 import { Markdown } from '../layout'
-import { useElection } from './Election'
 
 export const ElectionDescription = (props: Omit<ReactMarkdownProps, 'children' | 'node'> & ChakraProps) => {
   const styles = useStyleConfig('ElectionDescription', props)
