+++ packages/chakra-components/src/components/Organization/index.ts
@@ -2,6 +2,3 @@ export * from './Avatar'
 export * from './Description'
 export * from './Header'
 export * from './Name'
-export * from './Organization'
-// maybe not required, since everything should be internal, but exported just in case
-export * from './use-organization-reducer'
