+++ packages/chakra-components/src/index.ts
@@ -1,6 +1,4 @@
 export * from './client'
 export * from './components'
 export * from './environment'
-export * from './i18n'
 export * from './theme'
-export * from './utils'
