+++ packages/react-providers/.eslintrc.yml
@@ -1,0 +1,10 @@
+env:
+  browser: true
+  es2021: true
+extends:
+  - react-app
+  - plugin:prettier/recommended
+parserOptions:
+  ecmaVersion: latest
+  sourceType: module
+rules: {}
