+++ packages/react-providers/src/election/ElectionProvider.tsx
@@ -1,0 +1,26 @@
+import { PropsWithChildren, createContext, useContext } from 'react'
+import { ElectionProviderProps, useElectionProvider } from './use-election-provider'
+
+export type ElectionState = ReturnType<typeof useElectionProvider>
+
+export const ElectionContext = createContext<ElectionState | undefined>(undefined)
+
+export const useElection = () => {
+  const ctxt = useContext(ElectionContext)
+  if (!ctxt) {
+    throw new Error(
+      'useElection returned `undefined`, maybe you forgot to wrap the component within <ElectionProvider />?'
+    )
+  }
+
+  return ctxt
+}
+
+export type ElectionProviderComponentProps = ElectionProviderProps
+
+export const ElectionProvider = ({ children, ...rest }: PropsWithChildren<ElectionProviderComponentProps>) => {
+  const value = useElectionProvider(rest)
+
+  return <ElectionContext.Provider value={value}>{children}</ElectionContext.Provider>
+}
+ElectionProvider.displayName = 'ElectionProvider'
