+++ packages/react-providers/src/election/index.ts
@@ -1,0 +1,2 @@
+export * from './ElectionProvider'
+export type { ElectionProviderProps } from './use-election-provider'
