+++ packages/react-providers/src/i18n/locales.ts
@@ -1,0 +1,9 @@
+import { RecursivePartial } from '../types'
+
+export const locales = {
+  errors: {
+    unauthorized: 'Not authorized to vote',
+  },
+}
+
+export type Locale = RecursivePartial<typeof locales>
