+++ packages/react-providers/src/types.ts
@@ -1,0 +1,9 @@
+export type ErrorPayload = any | unknown
+
+export type RecursivePartial<T> = {
+  [P in keyof T]?: T[P] extends (infer U)[]
+    ? RecursivePartial<U>[]
+    : T[P] extends object
+    ? RecursivePartial<T[P]>
+    : T[P]
+}
