+++ packages/react-providers/tsconfig.json
@@ -1,0 +1,4 @@
+{
+  "extends": "tsconfig/react-library.json",
+  "include": ["src"]
+}
