+++ packages/tsup.config.ts
@@ -1,0 +1,7 @@
+import { defineConfig } from 'tsup'
+
+export default defineConfig({
+  format: ['cjs', 'esm'],
+  target: 'es2019',
+  sourcemap: true,
+})
