+++ templates/chakra/src/components/Account.tsx
@@ -4,13 +4,13 @@ import { BiChevronDown, BiLogOut } from 'react-icons/bi'
 import { useAccount, useDisconnect } from 'wagmi'
 
 import '@rainbow-me/rainbowkit/styles.css'
-import { useClient } from '@vocdoni/chakra-components'
+import { useClient } from '@vocdoni/react-providers'
 
 const addressTextOverflow = (address: string) =>
   `${address.substring(0, 6)}...${address.substring(address.length - 4, address.length)}`
 
 export const Account = () => {
-  const { clear, account, balance } = useClient()
+  const { clear } = useClient()
   const { isConnected, address } = useAccount()
   const { disconnect } = useDisconnect()
