+++ templates/chakra/src/elements/Vote.tsx
@@ -1,12 +1,7 @@
 import { Box, Card, CardBody, CardHeader, Flex } from '@chakra-ui/react'
 import { ConnectButton } from '@rainbow-me/rainbowkit'
-import {
-  Election,
-  OrganizationAvatar,
-  OrganizationDescription,
-  OrganizationName,
-  OrganizationProvider,
-} from '@vocdoni/chakra-components'
+import { Election, OrganizationAvatar, OrganizationDescription, OrganizationName } from '@vocdoni/chakra-components'
+import { OrganizationProvider } from '@vocdoni/react-providers'
 import { PublishedElection } from '@vocdoni/sdk'
 import { useLoaderData } from 'react-router-dom'
